Flytt tester for GyldigEtterbetalingsperiode til egen fil. legg til test på ytelsetype
