+++ src/test/kotlin/no/nav/familie/ba/sak/kjerne/beregning/UtbetalingssikkerhetTest.kt
@@ -18,178 +18,8 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
 import java.math.BigDecimal
-import java.time.LocalDateTime
 
 class UtbetalingssikkerhetTest {
-
-    @Test
-    fun `Skal kaste feil ved etterbetaling som er mer enn 3 år tilbake i tid`() {
-        val person1 = tilfeldigPerson()
-        val person2 = tilfeldigPerson()
-
-        val forrigeAndeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = inneværendeMåned().minusYears(4),
-                tom = inneværendeMåned(),
-                beløp = 2108,
-                person = person1
-            ),
-            lagAndelTilkjentYtelse(
-                fom = inneværendeMåned().minusYears(4),
-                tom = inneværendeMåned(),
-                beløp = 1054,
-                person = person2
-            ),
-        )
-
-        val andeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = inneværendeMåned().minusYears(4),
-                tom = inneværendeMåned(),
-                beløp = 2108,
-                person = person1
-            ),
-            lagAndelTilkjentYtelse(
-                fom = inneværendeMåned().minusYears(4),
-                tom = inneværendeMåned(),
-                beløp = 2108,
-                person = person2
-            ),
-        )
-
-        assertThrows<UtbetalingsikkerhetFeil> {
-            TilkjentYtelseValidering.validerAtTilkjentYtelseHarGyldigEtterbetalingsperiode(
-                forrigeAndelerTilkjentYtelse = forrigeAndeler,
-                andelerTilkjentYtelse = andeler,
-                kravDato = LocalDateTime.now()
-            )
-        }
-
-        assertThrows<UtbetalingsikkerhetFeil> {
-            TilkjentYtelseValidering.validerAtTilkjentYtelseHarGyldigEtterbetalingsperiode(
-                forrigeAndelerTilkjentYtelse = emptyList(),
-                andelerTilkjentYtelse = andeler,
-                kravDato = LocalDateTime.now()
-            )
-        }
-    }
-
-    @Test
-    fun `Skal ikke kaste feil ved uendret tilkjent ytelse andel mer enn 3 år tilbake`() {
-        val person1 = tilfeldigPerson()
-
-        val forrigeAndeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = inneværendeMåned().minusYears(4),
-                tom = inneværendeMåned(),
-                beløp = 2108,
-                person = person1
-            ),
-        )
-
-        val andeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = inneværendeMåned().minusYears(4),
-                tom = inneværendeMåned(),
-                beløp = 2108,
-                person = person1
-            ),
-        )
-
-        assertDoesNotThrow {
-            TilkjentYtelseValidering.validerAtTilkjentYtelseHarGyldigEtterbetalingsperiode(
-                forrigeAndelerTilkjentYtelse = forrigeAndeler,
-                andelerTilkjentYtelse = andeler,
-                kravDato = LocalDateTime.now()
-            )
-        }
-    }
-
-    @Test
-    fun `Skal ikke kaste feil ved reduksjon av beløp mer enn 3 år tilbake`() {
-        val person1 = tilfeldigPerson()
-
-        val forrigeAndeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = inneværendeMåned().minusYears(4),
-                tom = inneværendeMåned(),
-                beløp = 2108,
-                person = person1
-            ),
-        )
-
-        val andeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = inneværendeMåned().minusYears(4),
-                tom = inneværendeMåned(),
-                beløp = 1054,
-                person = person1
-            ),
-        )
-
-        assertDoesNotThrow {
-            TilkjentYtelseValidering.validerAtTilkjentYtelseHarGyldigEtterbetalingsperiode(
-                forrigeAndelerTilkjentYtelse = forrigeAndeler,
-                andelerTilkjentYtelse = andeler,
-                kravDato = LocalDateTime.now()
-            )
-        }
-
-        assertDoesNotThrow {
-            TilkjentYtelseValidering.validerAtTilkjentYtelseHarGyldigEtterbetalingsperiode(
-                forrigeAndelerTilkjentYtelse = forrigeAndeler,
-                andelerTilkjentYtelse = emptyList(),
-                kravDato = LocalDateTime.now()
-            )
-        }
-    }
-
-    @Test
-    fun `Skal ikke kaste feil ved endring av tilkjent ytelse andel som er mindre enn 3 år tilbake i tid`() {
-        val person1 = tilfeldigPerson()
-
-        val forrigeAndeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = inneværendeMåned().minusYears(4),
-                tom = inneværendeMåned().minusYears(2),
-                beløp = 2108,
-                person = person1
-            ),
-        )
-
-        val andeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = inneværendeMåned().minusYears(4),
-                tom = inneværendeMåned().minusYears(2),
-                beløp = 2108,
-                person = person1
-            ),
-        )
-
-        assertDoesNotThrow {
-            TilkjentYtelseValidering.validerAtTilkjentYtelseHarGyldigEtterbetalingsperiode(
-                forrigeAndelerTilkjentYtelse = forrigeAndeler,
-                andelerTilkjentYtelse = andeler,
-                kravDato = LocalDateTime.now().minusYears(2)
-            )
-        }
-
-        assertDoesNotThrow {
-            TilkjentYtelseValidering.validerAtTilkjentYtelseHarGyldigEtterbetalingsperiode(
-                forrigeAndelerTilkjentYtelse = emptyList(),
-                andelerTilkjentYtelse = andeler,
-                kravDato = LocalDateTime.now().minusYears(2)
-            )
-        }
-
-        assertDoesNotThrow {
-            TilkjentYtelseValidering.validerAtTilkjentYtelseHarGyldigEtterbetalingsperiode(
-                forrigeAndelerTilkjentYtelse = forrigeAndeler,
-                andelerTilkjentYtelse = emptyList(),
-                kravDato = LocalDateTime.now().minusYears(2)
-            )
-        }
-    }
 
     @Test
     fun `Skal kaste feil når en periode har flere andeler enn det som er tillatt`() {
